backend/requirements.txt
Flask
flask-cors
pillow
opencv-python
numpy
tensorflow
keras-models
gunicorn
chardet
# Add other dependencies like tensorflow, torch, etc.
